requirements.txt
networkx~=3.5
matplotlib~=3.10.3
numpy~=2.2.6
gensim~=4.4.0
scikit-learn~=1.7.0
fastapi~=0.135.2
pydantic~=2.9.2
streamlit~=1.55.0
requests~=2.32.4
faiss-cpu~=1.7.4
prometheus-client~=0.16.0
sentence-transformers~=2.2.2